moving fee tracking into its own service due to a need for regular prices for metrics
