+++ packages/services/src/lib.rs
@@ -2,6 +2,7 @@ pub mod block_bundler;
 pub mod block_committer;
 pub mod block_importer;
 pub mod cost_reporter;
+pub mod fee_tracker;
 pub mod health_reporter;
 pub mod state_committer;
 pub mod state_listener;
