+++ packages/services/tests/state_listener.rs
@@ -3,13 +3,15 @@ use std::time::Duration;
 use metrics::prometheus::IntGauge;
 use mockall::predicate::eq;
 use services::{
-    state_committer::port::l1::testing::ApiMockWFees,
     state_listener::{port::Storage, service::StateListener},
     types::{L1Height, L1Tx, TransactionResponse},
     Result, Runner, StateCommitter, StateCommitterConfig,
 };
 use test_case::test_case;
-use test_helpers::mocks::{self, l1::TxStatus};
+use test_helpers::{
+    mocks::{self, l1::TxStatus},
+    noop_fee_tracker,
+};
 
 #[tokio::test]
 async fn successful_finalized_tx() -> Result<()> {
@@ -446,7 +448,7 @@ async fn block_inclusion_of_replacement_leaves_no_pending_txs() -> Result<()> {
         .returning(|| Box::pin(async { Ok(0) }));
 
     let mut committer = StateCommitter::new(
-        ApiMockWFees::new(l1_mock),
+        l1_mock,
         mocks::fuel::latest_height_is(0),
         setup.db(),
         StateCommitterConfig {
@@ -454,6 +456,7 @@ async fn block_inclusion_of_replacement_leaves_no_pending_txs() -> Result<()> {
             ..Default::default()
         },
         test_clock.clone(),
+        noop_fee_tracker(),
     );
 
     // Orig tx
@@ -549,7 +552,7 @@ async fn finalized_replacement_tx_will_leave_no_pending_tx(
         .returning(|| Box::pin(async { Ok(0) }));
 
     let mut committer = StateCommitter::new(
-        ApiMockWFees::new(l1_mock),
+        l1_mock,
         mocks::fuel::latest_height_is(0),
         setup.db(),
         crate::StateCommitterConfig {
@@ -557,6 +560,7 @@ async fn finalized_replacement_tx_will_leave_no_pending_tx(
             ..Default::default()
         },
         test_clock.clone(),
+        noop_fee_tracker(),
     );
 
     // Orig tx
